refactor(rsvp_service): split rsvp_repository into per-table repositories (SCRUM-571)
